+++ man-roxygen/arg_parameters.R
@@ -1,0 +1,1 @@
+#' @param parameters (`list()`) Parameter space in irace format. See the function [irace::readParameters()].